# frontend/packages/shared/auth-openapi.json
{
    "components": {"schemas":{"api.AuthResponse":{"properties":{"authUserId":{"type":"string"},"enckey":{"$ref":"#/components/schemas/api.Enckey"},"mail":{"type":"string"},"token":{"type":"string"},"user":{"$ref":"#/components/schemas/api.FullUser"},"userId":{"type":"string"}},"type":"object"},"api.ChangePasswordRequestBody":{"properties":{"password":{"minLength":8,"type":"string"}},"required":["password"],"type":"object"},"api.ChangePasswordResponse":{"properties":{"token":{"type":"string"}},"type":"object"},"api.EmptyResponse":{"type":"object"},"api.Enckey":{"properties":{"createdAt":{"type":"string"},"id":{"type":"string"},"key":{"type":"string"},"salt":{"type":"string"}},"type":"object"},"api.EnckeyResponse":{"properties":{"enckey":{"$ref":"#/components/schemas/api.Enckey"}},"type":"object"},"api.ErrorResponse":{"properties":{"message":{"type":"string"}},"type":"object"},"api.FullUser":{"properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"level":{"type":"integer"},"mail":{"type":"string"}},"type":"object"},"api.JWKResponse":{"properties":{"publicKey":{"additionalProperties":{"type":"string"},"type":"object"}},"type":"object"},"api.JWKSResponse":{"properties":{"keys":{"items":{"additionalProperties":{"type":"string"},"type":"object"},"type":"array","uniqueItems":false}},"type":"object"},"api.LoginRequestBody":{"properties":{"mail":{"type":"string"},"password":{"type":"string"}},"required":["mail","password"],"type":"object"},"api.RegisterRequest":{"properties":{"mail":{"type":"string"},"password":{"type":"string"},"user":{"$ref":"#/components/schemas/api.FullUser"}},"type":"object"},"api.RegisterResponse":{"properties":{"mail":{"type":"string"},"token":{"type":"string"},"user":{"$ref":"#/components/schemas/api.FullUser"},"userId":{"type":"string"}},"type":"object"},"api.SaveEnckeyRequestBody":{"properties":{"key":{"type":"string"},"salt":{"type":"string"}},"required":["key","salt"],"type":"object"}},"securitySchemes":{"BearerAuth":{"in":"header","name":"Authorization","type":"apiKey"}}},
    "info": {"description":"Memoneo authentication and encrypted key API.","title":"Memoneo Auth API","version":"0.0.0"},
    "externalDocs": {"description":"","url":""},
    "paths": {"/.well-known/jwks.json":{"get":{"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.JWKSResponse"}}},"description":"OK"}},"summary":"Get JSON Web Key Set","tags":["auth"]}},"/auth":{"get":{"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.AuthResponse"}}},"description":"OK"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ErrorResponse"}}},"description":"Unauthorized"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ErrorResponse"}}},"description":"Internal Server Error"}},"security":[{"BearerAuth":[]}],"summary":"Check auth","tags":["auth"]}},"/auth/password":{"post":{"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ChangePasswordRequestBody"}}},"description":"New password","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ChangePasswordResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ErrorResponse"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ErrorResponse"}}},"description":"Unauthorized"}},"security":[{"BearerAuth":[]}],"summary":"Change password","tags":["auth"]}},"/enckey":{"get":{"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.EnckeyResponse"}}},"description":"OK"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ErrorResponse"}}},"description":"Unauthorized"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ErrorResponse"}}},"description":"Internal Server Error"}},"security":[{"BearerAuth":[]}],"summary":"Get encrypted key","tags":["enckey"]}},"/enckey/save":{"post":{"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.SaveEnckeyRequestBody"}}},"description":"Encrypted key","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.EmptyResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ErrorResponse"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ErrorResponse"}}},"description":"Unauthorized"}},"security":[{"BearerAuth":[]}],"summary":"Save encrypted key","tags":["enckey"]}},"/login":{"post":{"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.LoginRequestBody"}}},"description":"Login credentials","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.AuthResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ErrorResponse"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ErrorResponse"}}},"description":"Unauthorized"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Login","tags":["auth"]}},"/logout":{"post":{"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.EmptyResponse"}}},"description":"OK"}},"summary":"Logout","tags":["auth"]}},"/publickey":{"get":{"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.JWKResponse"}}},"description":"OK"}},"summary":"Get public key","tags":["auth"]}},"/register":{"post":{"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.RegisterRequest"}}},"description":"Registration data","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.RegisterResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ErrorResponse"}}},"description":"Bad Request"}},"summary":"Register","tags":["auth"]}},"/signup":{"post":{"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.RegisterRequest"}}},"description":"Registration data","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.RegisterResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/api.ErrorResponse"}}},"description":"Bad Request"}},"summary":"Register","tags":["auth"]}}},
    "openapi": "3.1.0",
    "servers": [
        {"url":"/"}
    ]
}